requirements.txt
argparse
pydantic
jinja2
pyyaml
requests
colored
tabulate
